+++ src/onegov.org/onegov/org/assets/js/ace-mode-form.js
@@ -18,8 +18,8 @@ var FormHighlightRules = function() {
                 regex: /(___|\*\*\*|\.\.\.|@@@|YYYY.MM.DD|YYYY.MM.DD HH:MM|HH:MM|[0-9]+[0-9\.]+[0-9]+)/
             },
             {
-                token: 'keyword',
-                regex: /[\[\(]{1}[x ]?[\]\)]{1}/
+                token: ['keyword', ''],
+                regex: /([\[\(]{1}[x ]?[\]\)]{1})(.*)$/
             },
             {
                 token: 'keyword.operator',